Some clean up

+++ src/matrix_tester.cc
@@ -22,7 +22,7 @@ int main (int argCnt, char *argVctr [])  {
     std::cout << "\nTesting the Symmetric Matrix ...\n" << std::endl;
 
     SDMatrix smat (8, 8);
-    double          count = 0.0;
+    int      count = 0;
 
     for (SDMatrix::size_type r = 0; r < 8; ++r)
         for (SDMatrix::size_type c = 0; c < 8; ++c)
@@ -55,7 +55,7 @@ int main (int argCnt, char *argVctr [])  {
 
     {
         DDMatrix dmat (8, 3);
-        double          count = 0.0;
+        int      count = 0;
 
         for (DDMatrix::size_type i = 0; i < 8; ++i)
             for (DDMatrix::size_type j = 0; j < 3; ++j)
@@ -123,7 +123,7 @@ int main (int argCnt, char *argVctr [])  {
         DDMatrix dmat (8, 3);
         DDMatrix one_mat (3, 8);
 
-        double  count = 0.0;
+        int  count = 0;
 
         for (DDMatrix::size_type i = 0; i < 8; ++i)
             for (DDMatrix::size_type j = 0; j < 3; ++j)  {
@@ -150,7 +150,7 @@ int main (int argCnt, char *argVctr [])  {
 
         DDMatrix dmat (3, 8);
 
-        double  count = 0.0;
+        int  count = 0;
 
         for (DDMatrix::size_type i = 0; i < 3; ++i)
             for (DDMatrix::size_type j = 0; j < 8; ++j)
@@ -189,7 +189,7 @@ int main (int argCnt, char *argVctr [])  {
 
         DDMatrix dmat (8, 8);
 
-        double  count = 0.0;
+        int  count = 0;
 
         for (DDMatrix::size_type i = 0; i < 8; ++i)
             for (DDMatrix::size_type j = 0; j < 8; ++j)
@@ -337,7 +337,7 @@ int main (int argCnt, char *argVctr [])  {
         DDMatrix dmat1 (8, 8);
         DDMatrix dmat2 (8, 8);
 
-        double  count = 0.0;
+        int  count = 0;
 
         for (DDMatrix::size_type i = 0; i < 8; ++i)
             for (DDMatrix::size_type j = 0; j < 8; ++j)  {
@@ -363,7 +363,7 @@ int main (int argCnt, char *argVctr [])  {
         std::cout << "After Addition(3):" << std::endl;
         dmat4.dump (std::cout) << std::endl;
 
-        count = 0.0;
+        count = 0;
         for (DDMatrix::size_type i = 0; i < 8; ++i)
             for (DDMatrix::size_type j = 0; j < 8; ++j)
                 dmat1 (i, j) = ++count;
@@ -379,7 +379,7 @@ int main (int argCnt, char *argVctr [])  {
         std::cout << "After dmat2 += dmat1;:" << std::endl;
         dmat2.dump (std::cout) << std::endl;
 
-        count = 0.0;
+        count = 0;
         for (DDMatrix::size_type i = 0; i < 8; ++i)
             for (DDMatrix::size_type j = 0; j < 8; ++j)
                 dmat1 (i, j) = ++count;
@@ -410,14 +410,14 @@ int main (int argCnt, char *argVctr [])  {
         {
             DDMatrix m1 (1, 4);
             DDMatrix m2 (4, 1);
-            double          count = 0.0;
+            int      count = 0;
 
             std::cout << "\t[1 X 4] * [4 X 1] = [1 X 1]\n" << std::endl;
             for (DDMatrix::size_type i = 0; i < 1; ++i)
                 for (DDMatrix::size_type j = 0; j < 4; ++j)
                     m1 (i, j) = ++count;
 
-            count = 0.0;
+            count = 0;
             for (DDMatrix::size_type i = 0; i < 4; ++i)
                 for (DDMatrix::size_type j = 0; j < 1; ++j)
                     m2 (i, j) = ++count;
@@ -434,14 +434,14 @@ int main (int argCnt, char *argVctr [])  {
         {
             DDMatrix m1 (3, 3);
             DDMatrix m2 (3, 3);
-            double          count = 0.0;
+            int      count = 0;
 
             std::cout << "\t[3 X 3] * [3 X 3] = [3 X 3]\n" << std::endl;
             for (DDMatrix::size_type i = 0; i < 3; ++i)
                 for (DDMatrix::size_type j = 0; j < 3; ++j)
                     m1 (i, j) = ++count;
 
-            count = 0.0;
+            count = 0;
             for (DDMatrix::size_type i = 0; i < 3; ++i)
                 for (DDMatrix::size_type j = 0; j < 3; ++j)
                     m2 (i, j) = ++count;
@@ -461,7 +461,7 @@ int main (int argCnt, char *argVctr [])  {
             DDMatrix m2 (2, 3);
             DDMatrix zero_m2 (2, 3);
             DDMatrix m3 (3, 3);
-            double          count = 0.0;
+            int      count = 0;
 
             std::cout << "\t[3 X 2] * [2 X 3] = [3 X 3]\n" << std::endl;
             for (DDMatrix::size_type i = 0; i < 3; ++i)
@@ -470,7 +470,7 @@ int main (int argCnt, char *argVctr [])  {
                     zero_m1 (i, j) = 0.0;
                 }
 
-            count = 0.0;
+            count = 0;
             for (DDMatrix::size_type i = 0; i < 2; ++i)
                 for (DDMatrix::size_type j = 0; j < 3; ++j)  {
                     m2 (i, j) = ++count;
@@ -498,7 +498,7 @@ int main (int argCnt, char *argVctr [])  {
             DDMatrix m5 (3, 2);
             DDMatrix m6 (2, 3);
 
-            count = 0.0;
+            count = 0;
             for (DDMatrix::size_type i = 0; i < 3; ++i)
                 for (DDMatrix::size_type j = 0; j < 2; ++j)
                     m5 (i, j) = ++count;
@@ -525,8 +525,7 @@ int main (int argCnt, char *argVctr [])  {
         DDMatrix dmat2 (8, 8);
         DDMatrix dmat3 (3, 8);
         DDMatrix dmat4 (3, 8);
-
-        double  count = 0.0;
+        int      count = 0;
 
         for (DDMatrix::size_type i = 0; i < 8; ++i)
             for (DDMatrix::size_type j = 0; j < 8; ++j)
@@ -535,7 +534,7 @@ int main (int argCnt, char *argVctr [])  {
             for (DDMatrix::size_type j = 0; j < 8; ++j)
                 dmat2 (i, j) = --count;
 
-        count = 0.0;
+        count = 0;
         for (DDMatrix::size_type i = 0; i < 3; ++i)
             for (DDMatrix::size_type j = 0; j < 8; ++j)
                 dmat3 (i, j) = ++count;
@@ -600,7 +599,7 @@ int main (int argCnt, char *argVctr [])  {
         const   int     pre = std::cout.precision (8);
         DDMatrix dmat1 (3, 3);
         DDMatrix dmat2 (3, 3);
-        double          count = 0.0;
+        int      count = 0;
 
         dmat1 (0, 0) = 2.0;
         dmat1 (0, 1) = 3.0;
@@ -693,7 +692,7 @@ int main (int argCnt, char *argVctr [])  {
 
         DDMatrix dmat2 (8, 3);
 
-        double  count = 0.0;
+        double  count = 0;
 
         for (DDMatrix::size_type i = 0; i < 8; ++i)
             for (DDMatrix::size_type j = 0; j < 3; ++j)
@@ -707,7 +706,7 @@ int main (int argCnt, char *argVctr [])  {
 
         DDMatrix dmat3 (8, 8);
 
-        count = 0.0;
+        count = 0;
         for (DDMatrix::size_type i = 0; i < 8; ++i)
             for (DDMatrix::size_type j = 0; j < 8; ++j)
                 dmat3 (i, j) = ::rand ();
@@ -728,8 +727,7 @@ int main (int argCnt, char *argVctr [])  {
         DDMatrix dmat (8, 4);
         DDMatrix null_mat (1, 4);
         DDMatrix two_mat (4, 1);
-
-        double  count = 0.0;
+        int      count = 0;
 
         for (DDMatrix::size_type i = 0; i < 8; ++i)
             for (DDMatrix::size_type j = 0; j < 4; ++j)
@@ -811,7 +809,7 @@ int main (int argCnt, char *argVctr [])  {
         std::cout << "\nTesting is_normal() method ...\n" << std::endl;
 
         DDMatrix dmat (3, 3);
-        double          count = 0.0;
+        int      count = 0;
 
         for (DDMatrix::size_type i = 0; i < 3; ++i)
             for (DDMatrix::size_type j = 0; j < 3; ++j)
@@ -851,7 +849,7 @@ int main (int argCnt, char *argVctr [])  {
                   << std::endl;
 
         SDMatrix dmat (10, 10);
-        double          count = 0.0;
+        int      count = 0;
 
         for (DDMatrix::size_type i = 0; i < 10; ++i)
             for (DDMatrix::size_type j = 0; j < 10; ++j)
@@ -873,7 +871,7 @@ int main (int argCnt, char *argVctr [])  {
 
         SDMatrix dmat2 (5, 5);
 
-        count = 0.0;
+        count = 0;
         for (DDMatrix::size_type i = 0; i < 5; ++i)
             for (DDMatrix::size_type j = 0; j < 5; ++j)
                 dmat2 (i, j) = ++count;
@@ -897,7 +895,7 @@ int main (int argCnt, char *argVctr [])  {
                   << std::endl;
 
         DDMatrix dmat (10, 10);
-        double          count = 0.0;
+        int      count = 0;
 
         for (DDMatrix::size_type i = 0; i < 10; ++i)
             for (DDMatrix::size_type j = 0; j < 10; ++j)
@@ -920,7 +918,7 @@ int main (int argCnt, char *argVctr [])  {
 
         DDMatrix dmat2 (5, 5);
 
-        count = 0.0;
+        count = 0;
         for (DDMatrix::size_type i = 0; i < 5; ++i)
             for (DDMatrix::size_type j = 0; j < 5; ++j)
                 dmat2 (i, j) = ++count;
@@ -946,7 +944,7 @@ int main (int argCnt, char *argVctr [])  {
 
         {
             DDMatrix dmat (8, 4);
-            double          count = 0.0;
+            int      count = 0;
 
             for (DDMatrix::size_type i = 0; i < 8; ++i)
                 for (DDMatrix::size_type j = 0; j < 4; ++j)
@@ -983,7 +981,7 @@ int main (int argCnt, char *argVctr [])  {
 
         {
             DDMatrix dmat (4, 8);
-            double          count = 0.0;
+            int      count = 0;
 
             for (DDMatrix::size_type i = 0; i < 4; ++i)
                 for (DDMatrix::size_type j = 0; j < 8; ++j)
@@ -1020,7 +1018,7 @@ int main (int argCnt, char *argVctr [])  {
 
         {
             DDMatrix dmat (8, 8);
-            double          count = 0.0;
+            int      count = 0;
 
             for (DDMatrix::size_type i = 0; i < 8; ++i)
                 for (DDMatrix::size_type j = 0; j < 8; ++j)
@@ -1065,7 +1063,7 @@ int main (int argCnt, char *argVctr [])  {
 
         {
             DDMatrix dmat (8, 4);
-            double          count = 0.0;
+            int      count = 0;
 
             for (DDMatrix::size_type i = 0; i < 8; ++i)
                 for (DDMatrix::size_type j = 0; j < 4; ++j)
@@ -1097,7 +1095,7 @@ int main (int argCnt, char *argVctr [])  {
 
         {
             DDMatrix dmat (4, 8);
-            double          count = 0.0;
+            int      count = 0;
 
             for (DDMatrix::size_type i = 0; i < 4; ++i)
                 for (DDMatrix::size_type j = 0; j < 8; ++j)
@@ -1129,7 +1127,7 @@ int main (int argCnt, char *argVctr [])  {
 
         {
             DDMatrix dmat (8, 8);
-            double          count = 0.0;
+            int      count = 0;
 
             for (DDMatrix::size_type i = 0; i < 8; ++i)
                 for (DDMatrix::size_type j = 0; j < 8; ++j)
@@ -1170,7 +1168,7 @@ int main (int argCnt, char *argVctr [])  {
         std::cout << "\nTesting lud() method ...\n" << std::endl;
 
         DDMatrix dmat (8, 8);
-        double          count = 0.0;
+        int      count = 0;
 
         for (DDMatrix::size_type i = 0; i < 8; ++i)
             for (DDMatrix::size_type j = 0; j < 4; ++j)
